Moved filter to shared #157
