+++ src/main/webapp/app/home/home-map/home-map.component.ts
@@ -1,11 +1,7 @@
 import {Component, HostListener, OnInit, ViewChild} from '@angular/core';
 import {AgmMap} from "@agm/core";
 import {MarkerService} from "../../entities/marker/marker.service";
-import {Marker} from "../../entities/marker/marker.model";
 import {JhiAlertService} from "ng-jhipster";
-import {FilterService} from '../../entities/filter/filter.service';
-import {Filter} from '../../entities/filter/filter.model';
-import {HttpResponse} from "@angular/common/http";
 
 @Component({
     selector: 'up-home-map',
