+++ src/main/webapp/app/shared/filter/filter.model.ts
