+++ src/main/webapp/app/shared/filter/filter.service.ts
@@ -2,7 +2,7 @@ import { HttpClient, HttpResponse } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 import { SERVER_API_URL } from '../../app.constants';
 import {Injectable} from '@angular/core';
-import {createRequestOption} from '../../shared';
+import {createRequestOption} from '../index';
 import {Filter} from './filter.model';
 
 export type EntityResponseType = HttpResponse<Filter>;
